requirements.txt
matplotlib
numpy
opencv_python
Pillow
streamlit
base64
BytesIO